Bayesian Paired comparison regression models in Stan

This is the main function of the package. This function utilizes precompiled stan models to sample the posterior distribution of the specified model with the input data. For more information and larger examples of usage see the vignettes.

bpc(
  data,
  player0,
  player1,
  player0_score = NULL,
  player1_score = NULL,
  result_column = NULL,
  z_player1 = NULL,
  cluster = NULL,
  predictors = NULL,
  model_type,
  solve_ties = "random",
  win_score = "higher",
  priors = NULL,
  chains = 4,
  iter = 2000,
  warmup = 1000,
  show_chain_messages = TRUE,
  seed = NA
)

Arguments

data
A data frame containing the observations. The other parameters specify the name of the columns

player0
A string with name of the column containing the players 0. This column should be of string/character type and not be of factor type.

player1
A string with name of the column containing the players 0. This column should be of string/character type and not be of factor type.

player0_score
A string with name of the column containing the scores of players 0

player1_score
A string with name of the column containing the scores of players 1

result_column
A string with name of the column containing the winners. 0 for player 0, 1 for player 1 and 2 for ties

z_player1
A string with the name of the column containing the order effect for player 1. E.g. if player1 has the home advantage this column should have 1 otherwise it should have 0

cluster
A string with the name of the column containing the cluster for the observation. To be used with a random effects model. This column should contain strings

predictors
A data frame that contains the players predictors values when using a generalized model. Only numeric values are accepted. Booleans are accepted but will be cast into integers. The first column should be for the player name, the others will be the predictors. The column names will be used as name for the predictors

model_type

'bt' for the Bradley Terry model. Ref: Bradley-Terry 1952,

'davidson' the Davidson model to handle for ties. Ref: Davidson 1970

'btordereffect' for the Bradley-Terry with order effect (for home advantage). Ref: Davidson 1977

'davidsonordereffect' for the Davidson model with order effect (for home advantage) and ties. Ref: Davidson 1977

'btgeneralized': for the generalized Bradley Terry model for subject specific predictors

'davidsongeneralized' for the generalized Davidson model for subject specific predictors

'btU': for the Bradley-Terry with random effects

'davidsonU': For Davidson model with random effects

solve_ties
A string for the method of handling ties.

'random' for converting ties randomly,

'remove' for removing the tie occurrencess

'none' to ignore ties. This requires a model capable of handling ties

win_score
A string that indicates if which score should win

'higher' score is winner

'lower' score is winner

priors
A list with the parameters for the priors.

'prior_lambda_mu' Mean value of the lambda parameter in the all models. For the generalized this is also the prior for the B the parameter for lambda ~ normal(mu, std)

'prior_lambda_std' Standard deviation of the lambda parameter in the all models. lambda ~ normal(mu, std)

'prior_nu_mu' Mean value of the nu parameter in the Davidson models. nu ~ normal(mu, std)

'prior_nu_std' Standard deviation ofnu parameter in the Davidson models. nu ~ normal(mu, std). Default = 0.3

'prior_gm_mu' Mean value of the gm in the ordered effect model. gm ~ normal(mu, std). Default = 0

'prior_gm_std' Standard deviation of the gm parameter in the ordered effect model. gm ~ normal(mu, std). Default =

'prior_U_std' Standard deviation of the U parameter in the random effects model. U ~ normal(0, std). Default = 3.0

chains
Number of chains passed to Stan sampling. Positive integer, default=4. For more information consult Stan documentation

iter
Number of iterations passed to Stan sampling. Positive integer, default =2000. For more information consult Stan documentation

warmup
Number of iteration for the warmup passed to Stan sampling. Positive integer, default 1000. For more information consult Stan documentation

show_chain_messages
Hide chain messages from Stan

seed
a random seed for Stan

Value

An object of the class bpc. This object should be used in conjunction with the several auxiliary functions from the package

References

Bradley RA, Terry ME (1952). “Rank Analysis of Incomplete Block Designs I: The Method of Paired Comparisons.” Biometrika, 39, 324–45. Davidson RR (1970). “On Extending the Bradley-Terry Model to Accommodate Ties in Paired Comparison Experiments.” Journal of the American Statistical Association, 65, 317–328. Davidson, Roger R., and Robert J. Beaver (1977). "On extending the Bradley-Terry model to incorporate within-pair order effects." Biometrics: 693-702. Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. http://mc-stan.org/.

Examples

#For the simple Bradley-Terry model
bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')

#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 4.6e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.46 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.096633 seconds (Warm-up)
#> Chain 1:                0.099019 seconds (Sampling)
#> Chain 1:                0.195652 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 1.5e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.100051 seconds (Warm-up)
#> Chain 2:                0.093553 seconds (Sampling)
#> Chain 2:                0.193604 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 1.7e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.17 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.103033 seconds (Warm-up)
#> Chain 3:                0.110924 seconds (Sampling)
#> Chain 3:                0.213957 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 1.5e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.15 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.10059 seconds (Warm-up)
#> Chain 4:                0.091725 seconds (Sampling)
#> Chain 4:                0.192315 seconds (Total)
#> Chain 4: 

#> Estimated baseline parameters with HPD intervals:
#> 
#> 
#> Parameter                Mean   HPD_lower   HPD_higher     n_eff    Rhat
#> --------------------  -------  ----------  -----------  --------  ------
#> lambda[Seles]           0.419      -2.490        3.277   811.021   1.006
#> lambda[Graf]            0.834      -1.955        3.681   790.160   1.005
#> lambda[Sabatini]       -0.442      -3.309        2.336   808.931   1.005
#> lambda[Navratilova]    -0.057      -3.028        2.713   793.332   1.005
#> lambda[Sanchez]        -1.215      -4.138        1.537   809.149   1.005
#> NOTES:
#> * A higher lambda indicates a higher team ability