Generate a ranking of the ability based on sampling the posterior distribution of the ranks.

To print this object you should remove the last column PosteriorRank since it contain the whole posterior distribution for each case

get_rank_of_players_posterior(bpc_object, n = 1000)

Arguments

bpc_object
a bpc object

n
Number of times we will sample the posterior

Value

a matrix containing the posterior distribution of the ranks

Examples

# \donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')

#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 6.8e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.68 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.60062 seconds (Warm-up)
#> Chain 1:                0.631342 seconds (Sampling)
#> Chain 1:                1.23196 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 6.2e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.62 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.600522 seconds (Warm-up)
#> Chain 2:                0.63592 seconds (Sampling)
#> Chain 2:                1.23644 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 6.3e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.63 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.637397 seconds (Warm-up)
#> Chain 3:                0.616819 seconds (Sampling)
#> Chain 3:                1.25422 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 6e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.6 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.646573 seconds (Warm-up)
#> Chain 4:                0.674011 seconds (Sampling)
#> Chain 4:                1.32058 seconds (Total)
#> Chain 4: 

rank_m<-get_rank_of_players_posterior(m,n=100)
print(rank_m)

#>        lambda[Seles] lambda[Graf] lambda[Sabatini] lambda[Navratilova]
#>   [1,]             3            1                4                   2
#>   [2,]             2            3                4                   1
#>   [3,]             2            1                4                   3
#>   [4,]             3            1                5                   2
#>   [5,]             2            1                3                   4
#>   [6,]             1            2                4                   3
#>   [7,]             2            1                4                   3
#>   [8,]             2            1                4                   3
#>   [9,]             2            1                4                   3
#>  [10,]             1            2                4                   3
#>  [11,]             3            1                5                   2
#>  [12,]             1            2                4                   3
#>  [13,]             1            2                4                   3
#>  [14,]             1            2                3                   4
#>  [15,]             2            1                3                   4
#>  [16,]             2            1                4                   3
#>  [17,]             2            1                4                   3
#>  [18,]             2            1                3                   4
#>  [19,]             4            1                3                   2
#>  [20,]             2            1                4                   3
#>  [21,]             2            1                5                   3
#>  [22,]             3            2                5                   1
#>  [23,]             4            1                2                   3
#>  [24,]             2            1                3                   4
#>  [25,]             4            1                2                   3
#>  [26,]             2            1                4                   3
#>  [27,]             2            1                4                   3
#>  [28,]             2            1                3                   5
#>  [29,]             1            2                3                   4
#>  [30,]             1            2                4                   3
#>  [31,]             2            1                4                   3
#>  [32,]             3            2                5                   1
#>  [33,]             2            1                3                   4
#>  [34,]             2            3                4                   1
#>  [35,]             2            1                4                   3
#>  [36,]             3            1                4                   2
#>  [37,]             3            2                4                   1
#>  [38,]             4            1                3                   2
#>  [39,]             2            1                5                   3
#>  [40,]             2            1                4                   3
#>  [41,]             1            2                4                   3
#>  [42,]             2            1                4                   3
#>  [43,]             2            1                3                   4
#>  [44,]             1            2                4                   3
#>  [45,]             2            1                4                   3
#>  [46,]             1            2                3                   4
#>  [47,]             2            1                4                   3
#>  [48,]             2            3                4                   1
#>  [49,]             4            1                2                   3
#>  [50,]             4            2                1                   3
#>  [51,]             5            1                3                   4
#>  [52,]             1            3                5                   2
#>  [53,]             1            2                4                   3
#>  [54,]             1            2                4                   3
#>  [55,]             2            1                4                   3
#>  [56,]             2            1                3                   4
#>  [57,]             3            1                4                   2
#>  [58,]             3            1                4                   2
#>  [59,]             1            2                4                   3
#>  [60,]             1            2                4                   3
#>  [61,]             2            1                3                   4
#>  [62,]             2            1                5                   3
#>  [63,]             3            1                4                   2
#>  [64,]             3            1                2                   4
#>  [65,]             3            1                4                   2
#>  [66,]             3            1                4                   2
#>  [67,]             1            3                5                   2
#>  [68,]             2            1                4                   3
#>  [69,]             1            2                4                   3
#>  [70,]             2            1                4                   3
#>  [71,]             1            2                4                   3
#>  [72,]             2            1                4                   3
#>  [73,]             2            1                3                   5
#>  [74,]             1            2                5                   4
#>  [75,]             2            1                3                   4
#>  [76,]             1            2                4                   3
#>  [77,]             2            1                4                   3
#>  [78,]             2            1                3                   4
#>  [79,]             3            1                5                   2
#>  [80,]             2            1                4                   3
#>  [81,]             2            1                4                   3
#>  [82,]             2            1                4                   3
#>  [83,]             2            1                5                   3
#>  [84,]             3            1                4                   2
#>  [85,]             2            1                5                   4
#>  [86,]             1            2                4                   3
#>  [87,]             2            1                4                   3
#>  [88,]             3            1                4                   2
#>  [89,]             2            1                4                   3
#>  [90,]             2            1                4                   3
#>  [91,]             2            1                4                   3
#>  [92,]             3            1                4                   2
#>  [93,]             3            1                4                   2
#>  [94,]             2            1                3                   4
#>  [95,]             1            2                4                   3
#>  [96,]             2            1                3                   4
#>  [97,]             2            1                3                   4
#>  [98,]             3            1                4                   2
#>  [99,]             3            1                4                   2
#> [100,]             1            2                5                   4
#>        lambda[Sanchez]
#>   [1,]               5
#>   [2,]               5
#>   [3,]               5
#>   [4,]               4
#>   [5,]               5
#>   [6,]               5
#>   [7,]               5
#>   [8,]               5
#>   [9,]               5
#>  [10,]               5
#>  [11,]               4
#>  [12,]               5
#>  [13,]               5
#>  [14,]               5
#>  [15,]               5
#>  [16,]               5
#>  [17,]               5
#>  [18,]               5
#>  [19,]               5
#>  [20,]               5
#>  [21,]               4
#>  [22,]               4
#>  [23,]               5
#>  [24,]               5
#>  [25,]               5
#>  [26,]               5
#>  [27,]               5
#>  [28,]               4
#>  [29,]               5
#>  [30,]               5
#>  [31,]               5
#>  [32,]               4
#>  [33,]               5
#>  [34,]               5
#>  [35,]               5
#>  [36,]               5
#>  [37,]               5
#>  [38,]               5
#>  [39,]               4
#>  [40,]               5
#>  [41,]               5
#>  [42,]               5
#>  [43,]               5
#>  [44,]               5
#>  [45,]               5
#>  [46,]               5
#>  [47,]               5
#>  [48,]               5
#>  [49,]               5
#>  [50,]               5
#>  [51,]               2
#>  [52,]               4
#>  [53,]               5
#>  [54,]               5
#>  [55,]               5
#>  [56,]               5
#>  [57,]               5
#>  [58,]               5
#>  [59,]               5
#>  [60,]               5
#>  [61,]               5
#>  [62,]               4
#>  [63,]               5
#>  [64,]               5
#>  [65,]               5
#>  [66,]               5
#>  [67,]               4
#>  [68,]               5
#>  [69,]               5
#>  [70,]               5
#>  [71,]               5
#>  [72,]               5
#>  [73,]               4
#>  [74,]               3
#>  [75,]               5
#>  [76,]               5
#>  [77,]               5
#>  [78,]               5
#>  [79,]               4
#>  [80,]               5
#>  [81,]               5
#>  [82,]               5
#>  [83,]               4
#>  [84,]               5
#>  [85,]               3
#>  [86,]               5
#>  [87,]               5
#>  [88,]               5
#>  [89,]               5
#>  [90,]               5
#>  [91,]               5
#>  [92,]               5
#>  [93,]               5
#>  [94,]               5
#>  [95,]               5
#>  [96,]               5
#>  [97,]               5
#>  [98,]               5
#>  [99,]               5
#> [100,]               3

# }