Get stanfit summary table of all parameters excluding log_lik.

Important to investigate the neff and the Rhat from the MCMC This excludes the log_lik paramter

get_stanfit_summary(bpc_object)

Arguments

bpc_object
a bpc object

Value

a data frame with the summary including quantiles, Rhat and neff

Examples

# \donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')

#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 5.1e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.51 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.413332 seconds (Warm-up)
#> Chain 1:                0.40345 seconds (Sampling)
#> Chain 1:                0.816782 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 4.4e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.44 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.45253 seconds (Warm-up)
#> Chain 2:                0.452311 seconds (Sampling)
#> Chain 2:                0.904841 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 4.5e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.45 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.43709 seconds (Warm-up)
#> Chain 3:                0.497875 seconds (Sampling)
#> Chain 3:                0.934965 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 4.4e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.44 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.438814 seconds (Warm-up)
#> Chain 4:                0.42148 seconds (Sampling)
#> Chain 4:                0.860294 seconds (Total)
#> Chain 4: 

s <- get_stanfit_summary(m)
print(s)

#>          Parameter        mean    se_mean       sd      2.5%         25%
#> 1  lambda_param[1]  0.49190075 0.06584384 1.466866 -2.366287 -0.49297226
#> 2  lambda_param[2]  0.92513564 0.06569247 1.436875 -1.841440 -0.04272316
#> 3  lambda_param[3] -0.36040995 0.06455815 1.453312 -3.185289 -1.33593251
#> 4  lambda_param[4]  0.01615506 0.06575449 1.442824 -2.854019 -0.94833208
#> 5  lambda_param[5] -1.14768436 0.06481527 1.447516 -3.943182 -2.11854317
#> 6        lambda[1]  0.49190075 0.06584384 1.466866 -2.366287 -0.49297226
#> 7        lambda[2]  0.92513564 0.06569247 1.436875 -1.841440 -0.04272316
#> 8        lambda[3] -0.36040995 0.06455815 1.453312 -3.185289 -1.33593251
#> 9        lambda[4]  0.01615506 0.06575449 1.442824 -2.854019 -0.94833208
#> 10       lambda[5] -1.14768436 0.06481527 1.447516 -3.943182 -2.11854317
#> 11              gm  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 12           U_std  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 13              nu  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 14            B[1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 15            B[2]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 16          U[1,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 17          U[2,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 18          U[3,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 19          U[4,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 20          U[5,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#>            50%        75%    97.5%    n_eff     Rhat
#> 1   0.46703261  1.4609051 3.464499 496.3072 1.007291
#> 2   0.91087827  1.8754772 3.783483 478.4173 1.008090
#> 3  -0.38079706  0.6192809 2.465597 506.7753 1.007527
#> 4  -0.02182563  0.9728277 2.896546 481.4771 1.006692
#> 5  -1.17231202 -0.1720603 1.764553 498.7604 1.008290
#> 6   0.46703261  1.4609051 3.464499 496.3072 1.007291
#> 7   0.91087827  1.8754772 3.783483 478.4173 1.008090
#> 8  -0.38079706  0.6192809 2.465597 506.7753 1.007527
#> 9  -0.02182563  0.9728277 2.896546 481.4771 1.006692
#> 10 -1.17231202 -0.1720603 1.764553 498.7604 1.008290
#> 11  0.00000000  0.0000000 0.000000      NaN      NaN
#> 12  0.00000000  0.0000000 0.000000      NaN      NaN
#> 13  0.00000000  0.0000000 0.000000      NaN      NaN
#> 14  0.00000000  0.0000000 0.000000      NaN      NaN
#> 15  0.00000000  0.0000000 0.000000      NaN      NaN
#> 16  0.00000000  0.0000000 0.000000      NaN      NaN
#> 17  0.00000000  0.0000000 0.000000      NaN      NaN
#> 18  0.00000000  0.0000000 0.000000      NaN      NaN
#> 19  0.00000000  0.0000000 0.000000      NaN      NaN
#> 20  0.00000000  0.0000000 0.000000      NaN      NaN

# }