Return the mean and the HPDI/credible intervals of the parameters of the model

Return a data frame with the mean and with high and low 95% hpd interval or credible interval for all parameters of the model

get_parameters(
  bpc_object,
  params = NULL,
  HPDI = TRUE,
  credMass = 0.95,
  n_eff = FALSE,
  Rhat = FALSE,
  keep_par_name = T
)

Arguments

bpc_object
a bpc object

params
a vector with the parameters to use. If null then all will be present

HPDI
should return the HPDI or the credible intervals. Default is returning the HPDI

credMass
probability mass for the summary stats

n_eff
Should include the number of effective samples in the df

Rhat
Should include the Rhat in the df

keep_par_name
keep the parameter name e.g. lambdaGraff instead of Graff. Default to T. Only valid for lambda, so we can have better ranks

Value

a data frame containing a column with the parameters, a column with mean and two columns with higher and lower intervals

Examples

# \donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')

#> Running MCMC with 4 parallel chains...
#> 
#> Chain 1 finished in 1.7 seconds.
#> Chain 2 finished in 1.8 seconds.
#> Chain 3 finished in 1.8 seconds.
#> Chain 4 finished in 1.7 seconds.
#> 
#> All 4 chains finished successfully.
#> Mean chain execution time: 1.7 seconds.
#> Total execution time: 2.0 seconds.

hpdi<-get_parameters(m)
print(hpdi)

#>             Parameter        Mean      Median HPD_lower HPD_higher
#> 1       lambda[Seles]  0.53113772  0.53787600  -2.23749    3.45011
#> 2        lambda[Graf]  0.96596751  0.99222650  -1.78792    3.79984
#> 3    lambda[Sabatini] -0.31953289 -0.31105600  -3.11668    2.55273
#> 4 lambda[Navratilova]  0.05819253  0.07323585  -2.67701    2.95873
#> 5     lambda[Sanchez] -1.10597649 -1.07783500  -4.07317    1.74432

# }