Get stanfit summary table of all parameters excluding log_lik.

Important to investigate the neff and the Rhat from the MCMC This excludes the log_lik paramter

get_stanfit_summary(bpc_object)

Arguments

bpc_object
a bpc object

Value

a data frame with the summary including quantiles, Rhat and neff

Examples

# \donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')

#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 1).
#> Chain 1: 
#> Chain 1: Gradient evaluation took 7.3e-05 seconds
#> Chain 1: 1000 transitions using 10 leapfrog steps per transition would take 0.73 seconds.
#> Chain 1: Adjust your expectations accordingly!
#> Chain 1: 
#> Chain 1: 
#> Chain 1: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 1: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 1: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 1: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 1: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 1: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 1: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 1: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 1: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 1: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 1: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 1: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 1: 
#> Chain 1:  Elapsed Time: 0.633526 seconds (Warm-up)
#> Chain 1:                0.668837 seconds (Sampling)
#> Chain 1:                1.30236 seconds (Total)
#> Chain 1: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 2).
#> Chain 2: 
#> Chain 2: Gradient evaluation took 6.6e-05 seconds
#> Chain 2: 1000 transitions using 10 leapfrog steps per transition would take 0.66 seconds.
#> Chain 2: Adjust your expectations accordingly!
#> Chain 2: 
#> Chain 2: 
#> Chain 2: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 2: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 2: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 2: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 2: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 2: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 2: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 2: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 2: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 2: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 2: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 2: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 2: 
#> Chain 2:  Elapsed Time: 0.623305 seconds (Warm-up)
#> Chain 2:                0.597921 seconds (Sampling)
#> Chain 2:                1.22123 seconds (Total)
#> Chain 2: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 3).
#> Chain 3: 
#> Chain 3: Gradient evaluation took 6.4e-05 seconds
#> Chain 3: 1000 transitions using 10 leapfrog steps per transition would take 0.64 seconds.
#> Chain 3: Adjust your expectations accordingly!
#> Chain 3: 
#> Chain 3: 
#> Chain 3: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 3: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 3: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 3: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 3: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 3: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 3: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 3: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 3: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 3: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 3: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 3: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 3: 
#> Chain 3:  Elapsed Time: 0.633007 seconds (Warm-up)
#> Chain 3:                0.629686 seconds (Sampling)
#> Chain 3:                1.26269 seconds (Total)
#> Chain 3: 
#> 
#> SAMPLING FOR MODEL 'bt' NOW (CHAIN 4).
#> Chain 4: 
#> Chain 4: Gradient evaluation took 7e-05 seconds
#> Chain 4: 1000 transitions using 10 leapfrog steps per transition would take 0.7 seconds.
#> Chain 4: Adjust your expectations accordingly!
#> Chain 4: 
#> Chain 4: 
#> Chain 4: Iteration:    1 / 2000 [  0%]  (Warmup)
#> Chain 4: Iteration:  200 / 2000 [ 10%]  (Warmup)
#> Chain 4: Iteration:  400 / 2000 [ 20%]  (Warmup)
#> Chain 4: Iteration:  600 / 2000 [ 30%]  (Warmup)
#> Chain 4: Iteration:  800 / 2000 [ 40%]  (Warmup)
#> Chain 4: Iteration: 1000 / 2000 [ 50%]  (Warmup)
#> Chain 4: Iteration: 1001 / 2000 [ 50%]  (Sampling)
#> Chain 4: Iteration: 1200 / 2000 [ 60%]  (Sampling)
#> Chain 4: Iteration: 1400 / 2000 [ 70%]  (Sampling)
#> Chain 4: Iteration: 1600 / 2000 [ 80%]  (Sampling)
#> Chain 4: Iteration: 1800 / 2000 [ 90%]  (Sampling)
#> Chain 4: Iteration: 2000 / 2000 [100%]  (Sampling)
#> Chain 4: 
#> Chain 4:  Elapsed Time: 0.64099 seconds (Warm-up)
#> Chain 4:                0.656549 seconds (Sampling)
#> Chain 4:                1.29754 seconds (Total)
#> Chain 4: 

#> Warning: Bulk Effective Samples Size (ESS) is too low, indicating posterior means and medians may be unreliable.
#> Running the chains for more iterations may help. See
#> http://mc-stan.org/misc/warnings.html#bulk-ess

s <- get_stanfit_summary(m)
print(s)

#>          Parameter        mean    se_mean       sd      2.5%         25%
#> 1  lambda_param[1]  0.42139212 0.06875923 1.449677 -2.507923 -0.53349602
#> 2  lambda_param[2]  0.85874669 0.06956294 1.409651 -1.954686 -0.05660934
#> 3  lambda_param[3] -0.41702225 0.06920432 1.425483 -3.265631 -1.36896676
#> 4  lambda_param[4] -0.03707412 0.06727246 1.437096 -2.879548 -1.00600939
#> 5  lambda_param[5] -1.19721049 0.06868723 1.427954 -4.000366 -2.16552413
#> 6        lambda[1]  0.42139212 0.06875923 1.449677 -2.507923 -0.53349602
#> 7        lambda[2]  0.85874669 0.06956294 1.409651 -1.954686 -0.05660934
#> 8        lambda[3] -0.41702225 0.06920432 1.425483 -3.265631 -1.36896676
#> 9        lambda[4] -0.03707412 0.06727246 1.437096 -2.879548 -1.00600939
#> 10       lambda[5] -1.19721049 0.06868723 1.427954 -4.000366 -2.16552413
#> 11              gm  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 12              nu  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 13            B[1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 14            B[2]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 15          U1_std  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 16          U2_std  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 17          U3_std  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 18          S[1,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 19          S[2,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 20          S[3,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 21          S[4,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 22          S[5,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 23         U1[1,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 24         U1[2,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 25         U1[3,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 26         U1[4,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 27         U1[5,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 28         U2[1,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 29         U2[2,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 30         U2[3,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 31         U2[4,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 32         U2[5,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 33         U3[1,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 34         U3[2,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 35         U3[3,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 36         U3[4,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#> 37         U3[5,1]  0.00000000        NaN 0.000000  0.000000  0.00000000
#>            50%        75%    97.5%    n_eff     Rhat
#> 1   0.43802835  1.4317158 3.209484 444.5092 1.012613
#> 2   0.87219208  1.8190977 3.517699 410.6460 1.013473
#> 3  -0.40877670  0.5670628 2.299090 424.2847 1.012354
#> 4  -0.01951915  0.9692795 2.666812 456.3486 1.013471
#> 5  -1.17101101 -0.1892750 1.524710 432.1916 1.013586
#> 6   0.43802835  1.4317158 3.209484 444.5092 1.012613
#> 7   0.87219208  1.8190977 3.517699 410.6460 1.013473
#> 8  -0.40877670  0.5670628 2.299090 424.2847 1.012354
#> 9  -0.01951915  0.9692795 2.666812 456.3486 1.013471
#> 10 -1.17101101 -0.1892750 1.524710 432.1916 1.013586
#> 11  0.00000000  0.0000000 0.000000      NaN      NaN
#> 12  0.00000000  0.0000000 0.000000      NaN      NaN
#> 13  0.00000000  0.0000000 0.000000      NaN      NaN
#> 14  0.00000000  0.0000000 0.000000      NaN      NaN
#> 15  0.00000000  0.0000000 0.000000      NaN      NaN
#> 16  0.00000000  0.0000000 0.000000      NaN      NaN
#> 17  0.00000000  0.0000000 0.000000      NaN      NaN
#> 18  0.00000000  0.0000000 0.000000      NaN      NaN
#> 19  0.00000000  0.0000000 0.000000      NaN      NaN
#> 20  0.00000000  0.0000000 0.000000      NaN      NaN
#> 21  0.00000000  0.0000000 0.000000      NaN      NaN
#> 22  0.00000000  0.0000000 0.000000      NaN      NaN
#> 23  0.00000000  0.0000000 0.000000      NaN      NaN
#> 24  0.00000000  0.0000000 0.000000      NaN      NaN
#> 25  0.00000000  0.0000000 0.000000      NaN      NaN
#> 26  0.00000000  0.0000000 0.000000      NaN      NaN
#> 27  0.00000000  0.0000000 0.000000      NaN      NaN
#> 28  0.00000000  0.0000000 0.000000      NaN      NaN
#> 29  0.00000000  0.0000000 0.000000      NaN      NaN
#> 30  0.00000000  0.0000000 0.000000      NaN      NaN
#> 31  0.00000000  0.0000000 0.000000      NaN      NaN
#> 32  0.00000000  0.0000000 0.000000      NaN      NaN
#> 33  0.00000000  0.0000000 0.000000      NaN      NaN
#> 34  0.00000000  0.0000000 0.000000      NaN      NaN
#> 35  0.00000000  0.0000000 0.000000      NaN      NaN
#> 36  0.00000000  0.0000000 0.000000      NaN      NaN
#> 37  0.00000000  0.0000000 0.000000      NaN      NaN

# }