Summary of the model bpc model.

Table 1: Contains the parameter estimates and respective HPD interval

Table 2: Contains the posterior probability for the combination of all players

Table 3: Contains the ranking of the players' abilities based on the posterior distribution of the ranks

# S3 method for bpc
summary(
  object,
  digits = 2,
  credMass = 0.95,
  HPDI = TRUE,
  show_probabilities = TRUE,
  ...
)

Arguments

object
bpc object

digits
number of decimal digits in the table

credMass
the probability mass for the credible interval

HPDI
True if show HPDI interval, False to show the credible (quantile) intervals

show_probabilities
should the tables of probabilities (Table 2) be displayed or not. Default to T but it is recommended to turn to F if either it has a large number of players (15+) or a large number of players and cluster, the table grows combinatorially.

...
additional parameters for the generic summary function. Not used.

Examples

# \donttest{
m<-bpc(data = tennis_agresti,
player0 = 'player0',
player1 = 'player1',
result_column = 'y',
model_type = 'bt',
solve_ties = 'none')
#> Running MCMC with 4 parallel chains...
#> 
#> Chain 1 finished in 2.4 seconds.
#> Chain 2 finished in 2.5 seconds.
#> Chain 3 finished in 2.6 seconds.
#> Chain 4 finished in 2.5 seconds.
#> 
#> All 4 chains finished successfully.
#> Mean chain execution time: 2.5 seconds.
#> Total execution time: 2.7 seconds.
summary(m)
#> Estimated baseline parameters with 95% HPD intervals:
#> 
#> Table: Parameters estimates
#> 
#> Parameter               Mean   Median   HPD_lower   HPD_higher
#> --------------------  ------  -------  ----------  -----------
#> lambda[Seles]           0.51     0.50       -2.07         3.36
#> lambda[Graf]            0.94     0.92       -1.81         3.56
#> lambda[Sabatini]       -0.35    -0.36       -3.05         2.33
#> lambda[Navratilova]     0.05     0.03       -2.71         2.72
#> lambda[Sanchez]        -1.12    -1.13       -3.82         1.71
#> NOTES:
#> * A higher lambda indicates a higher team ability
#> 
#> Posterior probabilities:
#> These probabilities are calculated from the predictive posterior distribution
#> for all player combinations
#> 
#> 
#> Table: Estimated posterior probabilites
#> 
#> i             j              i_beats_j   j_beats_i
#> ------------  ------------  ----------  ----------
#> Graf          Navratilova         0.70        0.30
#> Graf          Sabatini            0.67        0.33
#> Graf          Sanchez             0.86        0.14
#> Graf          Seles               0.63        0.37
#> Navratilova   Sabatini            0.62        0.38
#> Navratilova   Sanchez             0.78        0.22
#> Navratilova   Seles               0.40        0.60
#> Sabatini      Sanchez             0.66        0.34
#> Sabatini      Seles               0.34        0.66
#> Sanchez       Seles               0.21        0.79
#> 
#> Rank of the players' abilities:
#> The rank is based on the posterior rank distribution of the lambda parameter
#> 
#> Table: Estimated posterior ranks
#> 
#> Parameter      MedianRank   MeanRank   StdRank
#> ------------  -----------  ---------  --------
#> Graf                    1       1.40      0.63
#> Seles                   2       2.12      0.89
#> Navratilova             3       2.98      0.94
#> Sabatini                4       3.69      0.80
#> Sanchez                 5       4.80      0.49
# }